Wave_rover/requirements.txt
websocket-client>=1.3.0
requests>=2.27.0
pyautogui
keyboard
